Generate model outputs for calibration from a parameter set

calibration_out computes model outputs to be used for calibration routines.

calibration_out(v_params_calib, l_params_all, v_target_times = c(10, 20, 30))

Arguments

v_params_calib

Vector of parameters that need to be calibrated.

l_params_all

List with all parameters of the decision model.

v_target_times

Numeric vector with the model cycles (i.e., the times of the calibration targets) at which model outputs should be returned. Default = c(10, 20, 30), matching the times of SickSicker_targets.

Value

A list with Survival (Surv), Prevalence of Sick and Sicker (Prev), and proportion of Sicker (PropSicker) out of all sick (Sick+Sicker) individuals, each evaluated at v_target_times.

Examples

l_params_all <- load_all_params()
calibration_out(v_params_calib = c(p_S1S2 = 0.105, hr_S1 = 3, hr_S2 = 10),
                l_params_all   = l_params_all)
#> $Surv
#>        10        20        30 
#> 0.9771488 0.9185241 0.7722050 
#> 
#> $Prev
#>        10        20        30 
#> 0.3351821 0.4444411 0.4960935 
#> 
#> $PropSicker
#>        10        20        30 
#> 0.4906586 0.6792314 0.7399895 
#>